--- src/components/AboutProject/css/ClubCarousel.css
@@ -1,98 +1,0 @@
-.clubCarousel {
-    background-color: lightgrey;
-    width: 1400px;
-    border-radius: 8px;
-}
-.clubCarousel .club {
-    width:600px;
-    height:532px;
-    background-color: white;
-    padding:80px;
-
-}
-
-.clubCarousel .club-img img{
-    width:496px;
-    height:296px;
-    border-radius: 15px;
-}
-
-.clubCarousel .inf {
-    height:296px;
-    display:flex;
-    flex-direction: row;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-}
-
-.clubCarousel .club-inf {
-    margin-left: 54px;
-    overflow: hidden;
-    width: 568px;
-    height:100%;
-    box-sizing: border-box;
-}
-.clubCarousel .club-inf .club-name{
-    font-family: Open Sans;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 28px;
-    line-height: 36px;
-    color: #2D4C68;
-    margin-bottom: 16px;
-}
-.clubCarousel .club-inf .club-description{
-    font-family: Open Sans;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 18px;
-    line-height: 28px;
-    color: #2D4C68; 
-    opacity: 0.9;
-    text-overflow: ellipsis;
-}
-
-
-
-.clubCarousel .address-icon {
-    font-size: 1.5em;
-    color: #FFA62C;
-    margin-right: 11px;
-}
-.clubCarousel .arrow {
-    color: #FFA62C;
-    font-size: 1.5em;
-    margin-right: 20px;
-}
-
-.clubCarousel .arrow-icon {
-    height: 100%;
-}
-
-.clubCarousel .address {
-    display: flex;
-    margin-bottom: 20px;
-}
-
-.clubCarousel .address .text {
-    color: #2D4C68;
-    opacity: 0.6;
-    font-weight: 600;
-    vertical-align: center;
-}
-
-
-.clubCarousel .slick-dots li button,
-.clubCarousel .slick-dots li {
-    background: #C4C4C4;
-    width: 12px;
-    height: 12px;
-    border-radius: 100%;
-}
-
-.clubCarousel .slick-dots li.slick-active button,
-.clubCarousel .slick-dots li.slick-active {
-    background: #FA8C16;
-    width: 12px;
-}
