+++ src/service/AboutUsService.js
@@ -1,4 +1,4 @@
-import {BASE_URL} from "./config/ApiConfig";
+import {BASE_URL, DELETE_FILE_URL} from "./config/ApiConfig";
 import fetchRequest from "./FetchRequest";
 import {searchParameters} from "../context/SearchContext";
 
@@ -16,7 +16,7 @@ export const updateItemById = async (data) => {
             text: data.text,
             video: data.video,
             type: data.type,
-            number: 0
+            number: null
         })
         .then((response) => {
             return response.data;
@@ -25,3 +25,42 @@ export const updateItemById = async (data) => {
             return error.response.data;
         });
 };
+
+export const createItem = async (type) => {
+    return await fetchRequest
+        .post(BASE_URL + "/api/about_us_item/", {
+            picture: null,
+            text: null,
+            video: null,
+            type: type,
+            number: null
+        })
+        .then((response) => {
+            return response.data;
+        })
+        .catch((error) => {
+            return error.response.data;
+        });
+}
+
+export const deleteItem = async(id) => {
+    return await fetchRequest.delete(BASE_URL + "/api/about_us_item/" + id)
+        .then((response) => {
+            return response.data
+        }).catch((error) => {
+            return error.response.data
+        });
+}
+
+export const changeOrder = async (id, number) => {
+    return await fetchRequest
+        .patch(BASE_URL + "/api/about_us_item/" + id, {
+            number: number
+        })
+        .then((response) => {
+            return response.data;
+        })
+        .catch((error) => {
+            return error.response.data;
+        });
+}
